requirements.txt
Flask==1.1.4
Jinja2==2.11.3
Werkzeug==0.16.1
markupsafe==1.1.1
fluent-logger

